requirements.txt
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.21.0
pillow>=8.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
pyyaml>=5.4.0
requests>=2.26.0
pandas>=1.3.0
scipy>=1.7.0
streamlit>=1.12.0
plotly>=5.0.0
labelImg>=1.8.6
xmltodict>=0.12.0
flask>=2.0.0
flask-cors>=3.0.0
gradio>=3.0.0